requirements.txt
mediacloud==3.12.4
python-dotenv==0.19.1
pymongo==3.12.0
redis==3.5.3
celery==5.1.2
sqlalchemy==1.4.25
spacy==3.1.3
readability-lxml==0.8.1
beautifulsoup4==4.10.0
tldextract==3.1.2
dateparser==1.1.0
